backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
tortoise-orm==0.20.0
aiomysql==0.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
opencv-python-headless==4.9.0.80
numpy==1.26.3
